requirements.txt
Django==1.8
Pillow==2.8.1
SQLAlchemy==1.0.0
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==1.0.0
django-postgrespool==0.3.0
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.3.0
psycopg2==2.6
static3==0.5.1
wsgiref==0.1.2
